+++ package.json
@@ -1,5 +1,8 @@
-{
-  "version": "0.3.4-alpha.18",
+{ 
+  "name": "bonde-webpages",
+  "author": "Viviane",
+  "module": "dist/bonde-webpages.esm.js",
+  "version": "0.3.4",
   "license": "MIT",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
@@ -40,12 +43,10 @@
       "<rootDir>/src/setupTests.ts"
     ],
     "testPathIgnorePatterns": [
-      "<rootDir>/example/"
+      "<rootDir>/bonde-public/"
     ]
   },
-  "name": "bonde-webpages",
-  "author": "Viviane",
-  "module": "dist/bonde-webpages.esm.js",
+ 
   "devDependencies": {
     "@babel/core": "^7.9.0",
     "@storybook/addon-actions": "^5.3.18",
